Address password reset review feedback
